feat: Add regionModel import to server.js
This change imports regionModel from models to be used in the server.js file.

chore: Refactor regionAPI to handle root regions

This change modifies the regionAPI code to handle root regions by considering regions with "root" parent or without a parent as parentless.

chore: Modify default values in utils.js

This change modifies the default values in utils.js to set the parent region select and new parent ID select options to 'root' with the value '000000000000000000000000'.